+++ src/commonMain/kotlin/dev/inmo/sdi/Module.kt
@@ -1,0 +1,7 @@
+package dev.inmo.sdi
+
+import kotlinx.serialization.Contextual
+import kotlinx.serialization.Serializable
+
+@Serializable(ModuleFullSerializer::class)
+class Module internal constructor(base: Map<String, @Contextual Any>) : Map<String, Any> by base