+++ src/commonMain/kotlin/dev/inmo/sdi/utils/JsonAdaptations.kt
@@ -1,12 +1,18 @@
-package com.insanusmokrassar.sdi.utils
+package dev.inmo.sdi.utils
 
-import kotlinx.serialization.ImplicitReflectionSerializer
+import dev.inmo.sdi.getClassesForIncludingInSDI
 import kotlinx.serialization.InternalSerializationApi
 import kotlinx.serialization.json.*
 import kotlinx.serialization.modules.*
 import kotlin.reflect.KClass
 
 private typealias PackageOrOtherDependencyNamePair = Pair<String?, String?>
+
+private val namesToTheirClasses = getClassesForIncludingInSDI().flatMap {
+    (it.second + it.first.qualifiedName!!).map { name ->
+        name to it.first.qualifiedName!!
+    }
+}.toMap()
 
 private fun JsonElement.resolvePackageName(currentKey: String, otherDependenciesKeys: Set<String>): PackageOrOtherDependencyNamePair {
     return when (this) {
@@ -17,16 +23,20 @@ private fun JsonElement.resolvePackageName(currentKey: String, otherDependencies
                 it to null
             }
         } ?: throw IllegalArgumentException("Value on dependency name \"$currentKey\" is invalid: provided $this, but expected package name or other dependency name string")
-        is JsonObject -> return currentKey to null
-        is JsonArray -> return getPrimitive(0).contentOrNull ?.let { it to null } ?: throw IllegalArgumentException("Value on first argument of dependency value must be its package as a string, but was provided ${get(0)}")
+        is JsonObject -> if (currentKey in otherDependenciesKeys) {
+            null to currentKey
+        } else {
+            (namesToTheirClasses[currentKey] ?: currentKey) to null
+        }
+        is JsonArray -> return get(0).jsonPrimitive.contentOrNull ?.let { (namesToTheirClasses[it] ?: it) to null } ?: throw IllegalArgumentException("Value on first argument of dependency value must be its package as a string, but was provided ${get(0)}")
     }
 }
 
-@ImplicitReflectionSerializer
+@InternalSerializationApi
 internal fun createModuleBasedOnConfigRoot(
     jsonObject: JsonObject,
     moduleBuilder: (SerializersModuleBuilder.() -> Unit)? = null,
-    baseContext: SerialModule,
+    baseContext: SerializersModule,
     vararg additionalClassesToInclude: KClass<*>
 ): Json {
     lateinit var caches: Map<String, () -> Any>
@@ -53,7 +63,7 @@ internal fun createModuleBasedOnConfigRoot(
                 }
                 is JsonArray -> {
                     if (elemValue.size > 1) {
-                        elemValue.getObject(1)
+                        elemValue[1].jsonObject
                     } else {
                         JsonObject(emptyMap())
                     }
@@ -61,7 +71,7 @@ internal fun createModuleBasedOnConfigRoot(
             }
 
             val serializer = resolveSerializerByPackageName(packageName)
-            return@callback jsonStringFormat.fromJson(serializer, argumentsObject) as Any
+            return@callback jsonStringFormat.decodeFromJsonElement(serializer, argumentsObject) as Any
         }
     }.toMap()
 
@@ -105,10 +115,10 @@ internal fun createModuleBasedOnConfigRoot(
             }
         }
     )
-    return Json(
-        configuration = JsonConfiguration(useArrayPolymorphism = true),
-        context = context
-    ).also {
+    return Json {
+        useArrayPolymorphism = true
+        serializersModule = context
+    }.also {
         jsonStringFormat = it
     }
 }